feat: extract initialization helpers
